# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreetBite API - Street Vendor Order Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to phone number
      description: Sends a 6-digit OTP to the vendor's phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpRequest"
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendOtpResponse"
        "400":
          description: Invalid phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and login
      description: Verifies the OTP and returns authentication token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpRequest"
      responses:
        "200":
          description: OTP verified successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyOtpResponse"
        "400":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  
  # Vendors
...
  /vendors/{vendorId}/menu:
    get:
      operationId: getVendorMenu
      tags: [menu]
      summary: Get vendor menu
      description: Returns all menu items for a vendor
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [main, side, drink, dessert]
        - name: available
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Menu items retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create menu item
      description: Creates a new menu item (requires authentication)
      security:
        - bearerAuth: []
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemRequest"
      responses:
        "201":
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /menu/{menuItemId}:
    put:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update menu item
      description: Updates a menu item (requires authentication)
      security:
        - bearerAuth: []
      parameters:
        - name: menuItemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemRequest"
      responses:
        "200":
          description: Menu item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Menu item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete menu item
      description: Deletes a menu item (requires authentication)
      security:
        - bearerAuth: []
      parameters:
        - name: menuItemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Menu item deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Menu item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  
  # Orders
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    
    # Auth Schemas
    SendOtpRequest:
      type: object
      properties:
        phoneNumber:
          type: string
          pattern: '^\d{10}$'
          example: "9876543210"
      required:
        - phoneNumber
    SendOtpResponse:
      type: object
      properties:
        message:
          type: string
        expiresIn:
          type: integer
          description: OTP expiry time in seconds
      required:
        - message
        - expiresIn
    VerifyOtpRequest:
      type: object
      properties:
        phoneNumber:
          type: string
          pattern: '^\d{10}$'
        otp:
          type: string
          pattern: '^\d{6}$'
      required:
        - phoneNumber
        - otp
    VerifyOtpResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT authentication token
        vendor:
          $ref: "#/components/schemas/Vendor"
      required:
        - token
        - vendor
    
    # Vendor Schemas
    Vendor:
      type: object
      properties:
        id:
          type: integer
        phoneNumber:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        rating:
          type: number
          format: float
        isOpen:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - phoneNumber
        - name
        - description
        - category
        - rating
        - isOpen
...
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: integer
          description: Price in cents/paise
        category:
          type: string
          enum: [main, side, drink, dessert]
        emoji:
          type: string
        available:
          type: boolean
        rating:
          type: number
          format: float
        prepTime:
          type: integer
          description: Preparation time in minutes
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - vendorId
        - name
        - description
        - price
        - category
        - emoji
        - available
        - rating
        - prepTime
    CreateMenuItemRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        category:
          type: string
          enum: [main, side, drink, dessert]
        emoji:
          type: string
        available:
          type: boolean
        rating:
          type: number
          format: float
        prepTime:
          type: integer
      required:
        - name
        - description
        - price
        - category
        - emoji
        - prepTime
    UpdateMenuItemRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        category:
          type: string
          enum: [main, side, drink, dessert]
        emoji:
          type: string
        available:
          type: boolean
        rating:
          type: number
          format: float
        prepTime:
          type: integer
    
    # Order Schemas
...